Conflicts:
	scilab/modules/core/src/fortran/core_f.vfproj
	scilab/modules/elementary_functions/tests/nonreg_tests/bug_426.tst
	scilab/modules/fileio/src/c/mget.c
	scilab/modules/fileio/src/c/mput.c
	scilab/modules/overloading/macros/%hm_matrix.sci
	scilab/modules/overloading/macros/%i_b_s.sci
	scilab/modules/overloading/macros/%s_b_i.sci
	scilab/modules/overloading/macros/%s_b_s.sci
	scilab/modules/overloading/macros/generic_i_h.sci
	scilab/modules/ui_data/src/cpp/UpdateBrowseVar.cpp

+++ scilab/modules/fileio/src/cpp/mput.cpp
@@ -96,7 +96,7 @@ void C2F(mputnc) (int *fd, void * res, int *n1, char *type, int *ierr)
             MPUT_GEN_NC(int, c1);
             break;
         case 'l' :
-            MPUT_GEN_NC(int32_t, c1);
+            MPUT_GEN_NC(long long, c1);
             break;
         case 's' :
             MPUT_GEN_NC(short, c1);
@@ -117,7 +117,7 @@ void C2F(mputnc) (int *fd, void * res, int *n1, char *type, int *ierr)
                     MPUT_GEN_NC(unsigned int, c2);
                     break;
                 case 'l' :
-                    MPUT_GEN_NC(uint32_t, c2);
+                    MPUT_GEN_NC(unsigned long long, c2);
                     break;
                 case 's' :
                     MPUT_GEN_NC(unsigned short, c2);
@@ -198,7 +198,7 @@ void mput2 (FILE *fa, int swap, double *res, int n, char *type, int *ierr)
             MPUT_GEN(int, c1);
             break;
         case 'l' :
-            MPUT_GEN(int32_t, c1);
+            MPUT_GEN(long long, c1);
             break;
         case 's' :
             MPUT_GEN(short, c1);
@@ -219,7 +219,7 @@ void mput2 (FILE *fa, int swap, double *res, int n, char *type, int *ierr)
                     MPUT_GEN(unsigned int, c2);
                     break;
                 case 'l' :
-                    MPUT_GEN(uint32_t, c2);
+                    MPUT_GEN(unsigned long long, c2);
                     break;
                 case 's' :
                     MPUT_GEN(unsigned short, c2);
